chore(rust): refactor `Address`
