+++ implementations/rust/ockam/ockam_core/src/routing/address.rs
@@ -1,221 +1,13 @@
-use crate::access_control::IncomingAccessControl;
 use crate::compat::rand::{distributions::Standard, prelude::Distribution, random, Rng};
 use crate::compat::{
     string::{String, ToString},
-    sync::Arc,
     vec::Vec,
 };
-use crate::{debugger, DenyAll, OutgoingAccessControl, RelayMessage, Result};
-use core::cmp::Ordering;
+use crate::{AddressParseError, AddressParseErrorKind, Result, TransportType, LOCAL};
 use core::fmt::{self, Debug, Display};
 use core::ops::Deref;
 use core::str::from_utf8;
 use serde::{Deserialize, Serialize};
-
-/// A `Mailbox` controls the dispatch of incoming messages for a
-/// particular [`Address`]
-#[derive(Clone)]
-pub struct Mailbox {
-    address: Address,
-    incoming: Arc<dyn IncomingAccessControl>,
-    outgoing: Arc<dyn OutgoingAccessControl>,
-}
-
-impl Debug for Mailbox {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(
-            f,
-            "{} {{in:{:?} out:{:?}}}",
-            self.address, self.incoming, self.outgoing
-        )
-    }
-}
-
-impl Ord for Mailbox {
-    fn cmp(&self, rhs: &Self) -> Ordering {
-        self.address.cmp(&rhs.address)
-    }
-}
-impl PartialOrd for Mailbox {
-    fn partial_cmp(&self, rhs: &Self) -> Option<Ordering> {
-        Some(self.cmp(rhs))
-    }
-}
-impl PartialEq for Mailbox {
-    fn eq(&self, rhs: &Self) -> bool {
-        self.address == rhs.address
-    }
-}
-impl Eq for Mailbox {}
-
-impl Mailbox {
-    /// Create a new `Mailbox` with the given [`Address`], [`IncomingAccessControl`] and [`OutgoingAccessControl`]
-    pub fn new(
-        address: impl Into<Address>,
-        incoming: Arc<dyn IncomingAccessControl>,
-        outgoing: Arc<dyn OutgoingAccessControl>,
-    ) -> Self {
-        Self {
-            address: address.into(),
-            incoming,
-            outgoing,
-        }
-    }
-    /// Create a new `Mailbox` allowed to send and receive all messages
-    pub fn deny_all(address: impl Into<Address>) -> Self {
-        Self {
-            address: address.into(),
-            incoming: Arc::new(DenyAll),
-            outgoing: Arc::new(DenyAll),
-        }
-    }
-    /// Return a reference to the [`Address`] of this mailbox
-    pub fn address(&self) -> &Address {
-        &self.address
-    }
-    /// Return a reference to the [`IncomingAccessControl`] for this mailbox
-    pub fn incoming_access_control(&self) -> &Arc<dyn IncomingAccessControl> {
-        &self.incoming
-    }
-    /// Return a reference to the [`OutgoingAccessControl`] for this mailbox
-    pub fn outgoing_access_control(&self) -> &Arc<dyn OutgoingAccessControl> {
-        &self.outgoing
-    }
-}
-
-/// A collection of [`Mailbox`]es associated to given Context
-#[derive(Clone)]
-pub struct Mailboxes {
-    main_mailbox: Mailbox,
-    additional_mailboxes: Vec<Mailbox>,
-}
-
-impl Debug for Mailboxes {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(
-            f,
-            "{:?} + {:?}",
-            self.main_mailbox, self.additional_mailboxes
-        )
-    }
-}
-
-impl Mailboxes {
-    /// Create a new collection of `Mailboxes` from the given [`Mailbox`]es
-    pub fn new(main_mailbox: Mailbox, additional_mailboxes: Vec<Mailbox>) -> Self {
-        Self {
-            main_mailbox,
-            additional_mailboxes,
-        }
-    }
-
-    /// Create a new collection of `Mailboxes` for the given
-    /// [`Address`] with [`IncomingAccessControl`] and [`OutgoingAccessControl`]
-    pub fn main(
-        address: impl Into<Address>,
-        incoming_access_control: Arc<dyn IncomingAccessControl>,
-        outgoing_access_control: Arc<dyn OutgoingAccessControl>,
-    ) -> Self {
-        Self {
-            main_mailbox: Mailbox::new(
-                address.into(),
-                incoming_access_control,
-                outgoing_access_control,
-            ),
-            additional_mailboxes: vec![],
-        }
-    }
-
-    /// Return a list of [`Address`] containing all addresses represented by these `Mailboxes`
-    pub fn aliases(&self) -> Vec<Address> {
-        self.additional_mailboxes
-            .iter()
-            .map(|x| x.address.clone())
-            .collect()
-    }
-
-    /// Return the main [`Address`] for these `Mailboxes`
-    pub fn main_address(&self) -> Address {
-        self.main_mailbox.address.clone()
-    }
-
-    /// Return `true` if the given [`Address`] is represented by these `Mailboxes`
-    pub fn contains(&self, msg_addr: &Address) -> bool {
-        if &self.main_mailbox.address == msg_addr {
-            true
-        } else {
-            self.additional_mailboxes
-                .iter()
-                .any(|x| &x.address == msg_addr)
-        }
-    }
-
-    /// Return a reference to the [`Mailbox`] with the given [`Address`]
-    pub fn find_mailbox(&self, msg_addr: &Address) -> Option<&Mailbox> {
-        if &self.main_mailbox.address == msg_addr {
-            Some(&self.main_mailbox)
-        } else {
-            self.additional_mailboxes
-                .iter()
-                .find(|x| &x.address == msg_addr)
-        }
-    }
-
-    /// Return `true` if the given [`Address`] is authorized to post
-    /// the given [`RelayMessage`] to these `Mailboxes`
-    /// TODO docs are confusing
-    pub async fn is_incoming_authorized(&self, relay_msg: &RelayMessage) -> Result<bool> {
-        if let Some(mailbox) = self.find_mailbox(relay_msg.destination()) {
-            debugger::log_incoming_access_control(mailbox, relay_msg);
-
-            mailbox.incoming.is_authorized(relay_msg).await
-        } else {
-            warn!(
-                "Message from {} for {} does not match any addresses for this destination",
-                relay_msg.source(),
-                relay_msg.destination()
-            );
-            crate::deny()
-        }
-    }
-
-    /// Return `true` if these `Mailboxes` are authorized to post the
-    /// given [`RelayMessage`] to the given [`Address`]
-    /// TODO docs are confusing
-    pub async fn is_outgoing_authorized(&self, relay_msg: &RelayMessage) -> Result<bool> {
-        if let Some(mailbox) = self.find_mailbox(relay_msg.source()) {
-            debugger::log_outgoing_access_control(mailbox, relay_msg);
-
-            mailbox.outgoing.is_authorized(relay_msg).await
-        } else {
-            warn!(
-                "Message from {} for {} does not match any addresses for this origin",
-                relay_msg.source(),
-                relay_msg.destination()
-            );
-            crate::deny()
-        }
-    }
-
-    /// Return the list of [`Address`] represented by these `Mailboxes`
-    pub fn addresses(&self) -> Vec<Address> {
-        let mut addresses = vec![self.main_mailbox.address.clone()];
-        addresses.append(&mut self.aliases());
-        addresses
-    }
-
-    /// Return a reference to the main [`Mailbox`] for this collection
-    /// of `Mailboxes`
-    pub fn main_mailbox(&self) -> &Mailbox {
-        &self.main_mailbox
-    }
-
-    /// Return a reference to the additional [`Mailbox`]es for this
-    /// collection of `Mailboxes`
-    pub fn additional_mailboxes(&self) -> &Vec<Mailbox> {
-        &self.additional_mailboxes
-    }
-}
 
 /// A generic address type.
 ///
@@ -239,53 +31,9 @@ impl Mailboxes {
 #[derive(Serialize, Deserialize, Clone, Hash, Ord, PartialOrd, Eq, PartialEq)]
 pub struct Address {
     tt: TransportType,
+    // TODO: It's binary but in most cases we assume it to be an UTF-8 string
     inner: Vec<u8>,
 }
-
-/// An error which is returned when address parsing from string fails.
-#[derive(Debug)]
-pub struct AddressParseError {
-    kind: AddressParseErrorKind,
-}
-
-/// Enum to store the cause of an address parsing failure.
-#[derive(Debug, Clone, PartialEq, Eq)]
-#[non_exhaustive]
-pub enum AddressParseErrorKind {
-    /// Unable to parse address num in the address string.
-    InvalidType(core::num::ParseIntError),
-    /// Address string has more than one '#' separator.
-    MultipleSep,
-}
-
-impl AddressParseError {
-    /// Create new address parse error instance.
-    pub fn new(kind: AddressParseErrorKind) -> Self {
-        Self { kind }
-    }
-    /// Return the cause of the address parsing failure.
-    pub fn kind(&self) -> &AddressParseErrorKind {
-        &self.kind
-    }
-}
-
-impl Display for AddressParseError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match &self.kind {
-            AddressParseErrorKind::InvalidType(e) => {
-                write!(f, "Failed to parse address type: '{}'", e)
-            }
-            AddressParseErrorKind::MultipleSep => {
-                write!(
-                    f,
-                    "Invalid address string: more than one '#' separator found"
-                )
-            }
-        }
-    }
-}
-
-impl crate::compat::error::Error for AddressParseError {}
 
 impl Address {
     /// Creates a new address from separate transport type and data parts.
@@ -530,38 +278,6 @@ impl Distribution<Address> for Standard {
     fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Address {
         let address: [u8; 16] = rng.gen();
         hex::encode(address).as_bytes().into()
-    }
-}
-
-/// The transport type of an address.
-#[derive(Serialize, Deserialize, Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
-#[serde(transparent)]
-pub struct TransportType(u8);
-
-/// The local transport type.
-pub const LOCAL: TransportType = TransportType::new(0);
-
-impl TransportType {
-    /// Create a new transport type.
-    pub const fn new(n: u8) -> Self {
-        TransportType(n)
-    }
-
-    /// Is this the local transport type?
-    pub fn is_local(self) -> bool {
-        self == LOCAL
-    }
-}
-
-impl Display for TransportType {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-impl From<TransportType> for u8 {
-    fn from(ty: TransportType) -> Self {
-        ty.0
     }
 }
