+++ implementations/rust/ockam/ockam_core/src/routing/mod.rs
@@ -14,3 +14,9 @@ pub use message::*;
 
 mod macros;
 pub use macros::*;
+
+mod mailbox;
+pub use mailbox::*;
+
+mod transport_type;
+pub use transport_type::*;
