+++ implementations/rust/ockam/ockam_core/src/routing/transport_type.rs
@@ -1,0 +1,34 @@
+use core::fmt::{self, Debug, Display};
+use serde::{Deserialize, Serialize};
+
+/// The transport type of an [`Address`].
+#[derive(Serialize, Deserialize, Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
+#[serde(transparent)]
+pub struct TransportType(u8);
+
+/// The local transport type.
+pub const LOCAL: TransportType = TransportType::new(0);
+
+impl TransportType {
+    /// Create a new transport type.
+    pub const fn new(n: u8) -> Self {
+        TransportType(n)
+    }
+
+    /// Is this the local transport type?
+    pub fn is_local(self) -> bool {
+        self == LOCAL
+    }
+}
+
+impl Display for TransportType {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self.0)
+    }
+}
+
+impl From<TransportType> for u8 {
+    fn from(ty: TransportType) -> Self {
+        ty.0
+    }
+}
